keep original action names internally for accepting/declining shares
