+++ packages/web-pkg/src/composables/actions/files/index.ts
@@ -1,10 +1,10 @@
 export * from './useFileActions'
 export * from './useFileActionsSetReadme'
-export * from './useFileActionsSyncShare'
+export * from './useFileActionsAcceptShare'
 export * from './useFileActionsToggleHideShare'
 export * from './useFileActionsCopy'
 export * from './useFileActionsCreateQuicklink'
-export * from './useFileActionsUnsyncShare'
+export * from './useFileActionsDeclineShare'
 export * from './useFileActionsDelete'
 export * from './useFileActionsDownloadArchive'
 export * from './useFileActionsDownloadFile'
