+++ packages/web-pkg/src/composables/actions/files/useFileActionsDeclineShare.ts
@@ -1,9 +1,12 @@
 import { triggerShareAction } from '../../../helpers/share/triggerShareAction'
-
+import {
+  isLocationSharesActive,
+  isLocationSpacesActive,
+  createLocationShares
+} from '../../../router'
 import { Store } from 'vuex'
 import PQueue from 'p-queue'
 import { ShareStatus } from '@ownclouders/web-client/src/helpers/share'
-import { isLocationSharesActive, isLocationSpacesActive } from '../../../router'
 import { useCapabilityFilesSharingResharing, useCapabilityShareJailEnabled } from '../../capability'
 import { useClientService } from '../../clientService'
 import { useConfigurationManager } from '../../configuration'
@@ -12,9 +15,9 @@ import { useRouter } from '../../router'
 import { useStore } from '../../store'
 import { computed, unref } from 'vue'
 import { useGettext } from 'vue3-gettext'
-import { FileAction, FileActionOptions } from '../../actions'
+import { FileAction, FileActionOptions } from '../types'
 
-export const useFileActionsSyncShare = ({ store }: { store?: Store<any> } = {}) => {
+export const useFileActionsDeclineShare = ({ store }: { store?: Store<any> } = {}) => {
   store = store || useStore()
   const router = useRouter()
   const { $gettext, $ngettext } = useGettext()
@@ -35,8 +38,8 @@ export const useFileActionsSyncShare = ({ store }: { store?: Store<any> } = {}) 
           try {
             const share = await triggerShareAction({
               resource,
-              status: ShareStatus.accepted,
-              // Set hidden false here?
+              status: ShareStatus.declined,
+              // TODO: Hidden implications here?
               hasResharing: unref(hasResharing),
               hasShareJail: unref(hasShareJail),
               client: clientService.owncloudSdk,
@@ -56,16 +59,15 @@ export const useFileActionsSyncShare = ({ store }: { store?: Store<any> } = {}) 
     await Promise.all(triggerPromises)
 
     if (errors.length === 0) {
-      store.dispatch('Files/resetFileSelection')
-
       if (isLocationSpacesActive(router, 'files-spaces-generic')) {
         store.dispatch('showMessage', {
           title: $ngettext(
-            'Sync for the selected share was enabled successfully',
-            'Sync for the selected shares was enabled successfully',
+            'Sync for the selected share was disabled successfully',
+            'Sync for the selected shares was disabled successfully',
             resources.length
           )
         })
+        router.push(createLocationShares('files-shares-with-me'))
       }
 
       return
@@ -73,8 +75,8 @@ export const useFileActionsSyncShare = ({ store }: { store?: Store<any> } = {}) 
 
     store.dispatch('showErrorMessage', {
       title: $ngettext(
-        'Failed to enable sync for the the selected share',
-        'Failed to enable sync for the selected shares',
+        'Failed to disable sync for the the selected share',
+        'Failed to disable sync for the selected shares',
         resources.length
       ),
       errors
@@ -83,10 +85,10 @@ export const useFileActionsSyncShare = ({ store }: { store?: Store<any> } = {}) 
 
   const actions = computed((): FileAction[] => [
     {
-      name: 'sync-share',
-      icon: 'check',
+      name: 'decline-share',
+      icon: 'spam-3',
       handler: (args) => loadingService.addTask(() => handler(args)),
-      label: () => $gettext('Enable sync'),
+      label: () => $gettext('Disable sync'),
       isEnabled: ({ space, resources }) => {
         if (
           !isLocationSharesActive(router, 'files-shares-with-me') &&
@@ -100,18 +102,18 @@ export const useFileActionsSyncShare = ({ store }: { store?: Store<any> } = {}) 
 
         if (
           isLocationSpacesActive(router, 'files-spaces-generic') &&
-          (unref(space)?.driveType !== 'share' || resources.length > 1 || resources[0].path !== '/')
+          (space?.driveType !== 'share' || resources.length > 1 || resources[0].path !== '/')
         ) {
           return false
         }
 
-        const acceptDisabled = resources.some((resource) => {
-          return resource.status === ShareStatus.accepted
+        const declineDisabled = resources.some((resource) => {
+          return resource.status === ShareStatus.declined
         })
-        return !acceptDisabled
+        return !declineDisabled
       },
       componentType: 'button',
-      class: 'oc-files-actions-sync-share-trigger'
+      class: 'oc-files-actions-decline-share-trigger'
     }
   ])
